UnsortedArrayPriorityQueue / Created IsEmpty and Head Methods.
The Head method will peek at the array index with the highest priority. The method checks if the array is empty, and throws an exception error if so. It then loops through the array checking each index value, against each other, for the highest priority value. The highest priority index is then stored and used to return the <T> Item name from the array Index