UnsortedArrayPriorityQueue / Added ToString method and functionality
Method to override the ToString method and throw an exception if the array is empty (tailIndex is less than 0). Adds a bracket infront and after each returned array index value, seperated by an added comma in the nested loop. Returns each index value in the array, the Item and Priority.