UnsortedLinkedPriorityQueue / Added Remove method with code commenting
Method tracks the currentNode, highestPriorityNode and previousNode. Method compares the values of the Priorities for each Node in the LinkedList and keeps a track of the Highest Priority Node and the Previous Node. Removes the Highest Priority Node by changing the pointers of the Previous Nodes to the Node Next to the Highest Priority Node.