UnsortedLinkedPriorityQueue / Added Second Remove Method
Remove method traverses each Link keeping track of the count of the Nodes. The Highest Priority Node is then recorded. We Traverse the LinkedList a second time to find the Node previous to the HighestPriorityNode. Finally we update the pointers on the PreviousNode to point to the Node after the HighestPriorityNode. This skips over the HighestPriorityNode and removes it from the LinkedList.

This is extremely inefficient as the time complexity is O(n)

The second Remove method was created due to logic problems with the first Remove method.