SortedLinkedPriorityQueue / Added Add method with code commenting
Starting at the _head Node and while there is a next Node and while the next Nodes Priority is greater or equal to the newNodes Priority, traverse the LinkedList. Stop when the next Node is less than the newNodes Priority. Update the currentNode to this traversed Node. Point the newNode next pointer to the currentNodes Next (lower priority Node) pointer. Point the currentNodes next pointer to the newNode. We have now inserted the newNode in the correct position while maintaining the Priority Queue.