TEST Branch - UnsortedArrayPriorityQueue / Added Add_AddingItemWithNullNameDoesNotGetAddedToQueue  *Sorted and Unsorted Array Priority Queue*
Modified the Sorted and Unsorted version of this Test.

The test should check weather the Null name Item has been added to the queue. The Head method should return the Item and the assert should check that it is NOT the head of the Queue. If the Item is not the head, it means it has not been added to the queue, since there is only 1 item being added